+++ src/components/assistant-ui/custom-markdown/index.ts
@@ -1,3 +1,2 @@
-// Custom Markdown Placeholder Components
-// These are the components used by the integrated markdown system in markdown-text.tsx
-export { DocPlaceholder, LinkPlaceholder } from './placeholders';
+export { CustomDocReference } from "./CustomDocReference";
+export { CustomLinkReference } from "./CustomLinkReference";
